Выполняет ДЗ номер 5
